+++ bracket-bevy/src/consoles/virtual_console.rs
@@ -93,7 +93,7 @@ impl VirtualConsole {
     }
 
     fn at(&self, x: i32, y: i32) -> usize {
-        if let Ok(pos) = (((self.height as i32 - 1 - y) * self.width as i32) + x).try_into() {
+        if let Ok(pos) = (((self.height - 1 - y) * self.width) + x).try_into() {
             pos
         } else {
             0
@@ -172,7 +172,7 @@ impl ConsoleFrontEnd for VirtualConsole {
 
     fn print_centered(&mut self, y: i32, text: &str) {
         self.print(
-            (self.width as i32 / 2) - (text.to_string().len() / 2) as i32,
+            (self.width / 2) - (text.to_string().len() / 2) as i32,
             y,
             text,
         );
@@ -184,7 +184,7 @@ impl ConsoleFrontEnd for VirtualConsole {
 
     fn print_color_centered(&mut self, y: i32, fg: RGBA, bg: RGBA, text: &str) {
         self.print_color(
-            (self.width as i32 / 2) - (text.to_string().len() / 2) as i32,
+            (self.width / 2) - (text.to_string().len() / 2) as i32,
             y,
             text,
             fg,
