+++ bracket-terminal/src/consoles/flexible_console.rs
@@ -10,6 +10,7 @@ use ultraviolet::Vec2;
 use std::any::Any;
 
 /// Internal storage structure for sparse tiles.
+#[derive(Debug)]
 pub struct FlexiTile {
     pub position: PointF,
     pub z_order: i32,
@@ -22,6 +23,7 @@ pub struct FlexiTile {
 
 /// A sparse console. Rather than storing every cell on the screen, it stores just cells that have
 /// data.
+#[derive(Debug)]
 pub struct FlexiConsole {
     pub width: u32,
     pub height: u32,
@@ -378,7 +380,7 @@ impl Console for FlexiConsole {
         for c in &self.tiles {
             let x = c.position.x as usize;
             let y = c.position.y as usize;
-            let cell = layer.get_mut(x as usize, y as usize).unwrap();
+            let cell = layer.get_mut(x, y).unwrap();
             cell.ch = u32::from(c.glyph);
             cell.fg = XpColor::from(c.fg);
             cell.bg = XpColor::from(c.bg);
