+++ bracket-terminal/src/consoles/sparse_console.rs
@@ -7,7 +7,7 @@ use bracket_geometry::prelude::Rect;
 use std::any::Any;
 
 /// Internal storage structure for sparse tiles.
-#[derive(Clone, Copy, PartialEq)]
+#[derive(Debug, Clone, Copy, PartialEq)]
 pub struct SparseTile {
     pub idx: usize,
     pub glyph: FontCharType,
@@ -17,6 +17,7 @@ pub struct SparseTile {
 
 /// A sparse console. Rather than storing every cell on the screen, it stores just cells that have
 /// data.
+#[derive(Debug)]
 pub struct SparseConsole {
     pub width: u32,
     pub height: u32,
@@ -341,7 +342,7 @@ impl Console for SparseConsole {
         for c in &self.tiles {
             let x = c.idx % self.width as usize;
             let y = c.idx / self.width as usize;
-            let cell = layer.get_mut(x as usize, y as usize).unwrap();
+            let cell = layer.get_mut(x, y).unwrap();
             cell.ch = u32::from(c.glyph);
             cell.fg = c.fg.into();
             cell.bg = c.bg.into();
